Setup Radiator Classes and Basic Analysis Code
The radiator classes for each type all inherit from a base Radiator class so that arrays of them can be created and they can share common functions making the code cleaner and simpler to use. Also created a function where arrays of Radiator objects can be created using a table on input parameters.